+++ paimon-common/src/main/java/org/apache/paimon/utils/RoaringBitmap32.java
@@ -19,6 +19,7 @@
 package org.apache.paimon.utils;
 
 import org.apache.paimon.annotation.VisibleForTesting;
+import org.apache.paimon.predicate.LimitDirection;
 
 import org.roaringbitmap.IntIterator;
 import org.roaringbitmap.RoaringBitmap;
@@ -29,6 +30,8 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.Iterator;
 import java.util.Objects;
+
+import static org.apache.paimon.predicate.LimitDirection.HEAD;
 
 /** A compressed bitmap for 32-bit integer. */
 public class RoaringBitmap32 {
@@ -107,16 +110,16 @@ public class RoaringBitmap32 {
     }
 
     public RoaringBitmap32 limit(int k) {
-        return limit(k, true);
+        return limit(k, HEAD);
     }
 
-    public RoaringBitmap32 limit(int k, boolean head) {
+    public RoaringBitmap32 limit(int k, LimitDirection direction) {
         if (getCardinality() <= k) {
             return clone();
         }
 
         // limit from head
-        if (head) {
+        if (HEAD.equals(direction)) {
             return new RoaringBitmap32(roaringBitmap.limit(k));
         }
